first compile

+++ src/internal/NeoEsp32I2sXMethod.cpp
@@ -1,8 +1,9 @@
 #include <Arduino.h>
+#include "NeoPixelBus.h"
 #include "NeoEsp32I2sXMethod.h"
 
 uint32_t NeoEsp32I2sX8BusZero::s_i2sBufferSize = 0;
-uint8_t* NeoEsp32I2sX8BusZero::s_i2sBuffer = nullptr;
+uint32_t* NeoEsp32I2sX8BusZero::s_i2sBuffer = nullptr;
 
 size_t NeoEsp32I2sX8BusZero::s_MaxBusDataSize = 0;
 uint8_t NeoEsp32I2sX8BusZero::s_UpdateMap = 0;
